+++ hooks/useDownloadsPage.ts
@@ -1,15 +1,13 @@
-import React, { useCallback, useEffect, useState } from 'react'
-import { getDownloads } from '..';
+import { useCallback, useEffect, useState } from "react";
+import { getDownloads } from "../pages/downloads";
 
-const useDownloadsPage = () => {
-    // mapear click de botones con el slide y detener el carrousel
-  // mapear flechas izquierda y derecha para pasar de slide
+export const useDownloadsPage = () => {
   const [activeResource, setActiveResource] = useState<string>();
   const [downloads, setDownloads] = useState([]);
 
   useEffect(() => {
     getDownloads().then((downloads) => setDownloads(downloads));
-  }, [])
+  }, []);
 
   const handleClick = useCallback(
     (event: React.SyntheticEvent<HTMLButtonElement>) => {
@@ -21,12 +19,5 @@ const useDownloadsPage = () => {
     [setActiveResource]
   );
 
-  return {
-    activeResource,
-    setActiveResource,
-    downloads,
-    handleClick,
-  }
+  return { activeResource, downloads, handleClick }
 }
-
-export default useDownloadsPage
